tests/wpt/tests/css/css-transforms/transform-important.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Transforms: !important flag parsing</title>
    <link rel="help" href="https://crbug.com/340100643">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="target" style="transform: translateX(10px) !important; will-change: transform !important">Some text.</div>
  </body>
  <script>
test(() => {
  assert_equals(target.style.transform, 'translateX(10px)');
  assert_equals(target.style.willChange, 'transform');
});
  </script>
</html>
